offer percentage

+++ Application/Interfaces/ILocationService.cs
@@ -10,8 +10,8 @@ namespace Application.Interfaces
     public interface ILocationService
     {
         List<Locations> GetAllUserLocations(int userid);
-        int AddLocation(int userid, string Address, int Postalcode, string Username, string Mobile, string Email, string CityName, string StateName);
-        void AddLocationForUser(int userid, int potalcode, string Address, string Username, string Mobile, string Email, string CityName, string StateName);
+        int AddLocation(int userid, string Address, ulong Postalcode, string Username, string Mobile, string Email, string CityName, string StateName);
+        void AddLocationForUser(int userid, ulong potalcode, string Address, string Username, string Mobile, string Email, string CityName, string StateName);
         Locations GetLocationByLocationID(int LocationID);
         void UpdateLocation(Locations locations);
         void DeleteLocation(Locations locations);
